Refactor Parameter Configuration Structure in Backtesting Module (MINOR) (#830)
