+++ src/main/java/com/verlumen/tradestream/backtesting/GenotypeConverterImpl.java
@@ -3,8 +3,6 @@ package com.verlumen.tradestream.backtesting;
 import com.google.common.collect.ImmutableList;
 import com.google.inject.Inject;
 import com.google.protobuf.Any;
-import com.verlumen.tradestream.backtesting.params.ParamConfig;
-import com.verlumen.tradestream.backtesting.ParamConfigManager;
 import com.verlumen.tradestream.strategies.StrategyType;
 import io.jenetics.DoubleGene;
 import io.jenetics.Genotype;
