+++ src/main/java/com/verlumen/tradestream/backtesting/ParamConfigManager.java
@@ -1,6 +1,5 @@
 package com.verlumen.tradestream.backtesting;
 
-import com.verlumen.tradestream.backtesting.params.ParamConfig;
 import com.verlumen.tradestream.strategies.StrategyType;
 import java.io.Serializable;
