+++ src/main/java/com/verlumen/tradestream/backtesting/oscillators/BUILD
@@ -3,8 +3,8 @@ load("@rules_java//java:defs.bzl", "java_library")
 package(default_visibility = ["//visibility:public"])
 
 java_library(
-    name = "params_lib",
-    srcs = glob(["*.java"]),
+    name = "sma_rsi_param_config",
+    srcs = ["SmaRsiParamConfig.java"],
     deps = [
         "//protos:strategies_java_proto",
         "//src/main/java/com/verlumen/tradestream/backtesting:chromosome_spec",
