+++ src/test/java/com/verlumen/tradestream/backtesting/BUILD
@@ -57,7 +57,6 @@ java_test(
         "//src/main/java/com/verlumen/tradestream/backtesting:ga_engine_factory_impl",
         "//src/main/java/com/verlumen/tradestream/backtesting:param_config",
         "//src/main/java/com/verlumen/tradestream/backtesting:param_config_manager",
-        "//src/main/java/com/verlumen/tradestream/backtesting/params:params_lib",
         "//third_party:guava",
         "//third_party:guice",
         "//third_party:guice_testlib",
@@ -104,7 +103,6 @@ java_test(
         "//src/main/java/com/verlumen/tradestream/backtesting:genotype_converter_impl",
         "//src/main/java/com/verlumen/tradestream/backtesting:param_config",
         "//src/main/java/com/verlumen/tradestream/backtesting:param_config_manager",
-        "//src/main/java/com/verlumen/tradestream/backtesting/params:params_lib",
         "//third_party:guava",
         "//third_party:guice",
         "//third_party:guice_testlib",
