+++ src/test/java/com/verlumen/tradestream/backtesting/GenotypeConverterImplTest.java
@@ -11,9 +11,6 @@ import com.google.inject.testing.fieldbinder.Bind;
 import com.google.inject.testing.fieldbinder.BoundFieldModule;
 import com.google.protobuf.Any;
 import com.google.protobuf.Int32Value;
-import com.verlumen.tradestream.backtesting.ChromosomeSpec;
-import com.verlumen.tradestream.backtesting.params.ParamConfig;
-import com.verlumen.tradestream.backtesting.ParamConfigManager;
 import com.verlumen.tradestream.strategies.StrategyType;
 import io.jenetics.Chromosome;
 import io.jenetics.DoubleChromosome;
