+++ src/test/java/com/verlumen/tradestream/backtesting/oscillators/BUILD
@@ -7,7 +7,7 @@ java_test(
     deps = [
         "//protos:strategies_java_proto",
         "//src/main/java/com/verlumen/tradestream/backtesting:chromosome_spec",
-        "//src/main/java/com/verlumen/tradestream/backtesting/params:params_lib",
+        "//src/main/java/com/verlumen/tradestream/backtesting/oscillators:sma_rsi_param_config",
         "//third_party:guava",
         "//third_party:guice",
         "//third_party:guice_testlib",
