+++ src/test/java/com/verlumen/tradestream/backtesting/oscillators/SmaRsiParamConfigTest.java
@@ -1,4 +1,4 @@
-package com.verlumen.tradestream.backtesting.params;
+package com.verlumen.tradestream.backtesting.oscillators;
 
 import static com.google.common.truth.Truth.assertThat;
